+++ src/bda/portlet/sitenavigation/sitenavigation.py
@@ -1,3 +1,4 @@
+from Acquisition import aq_base
 from Acquisition import aq_inner
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -8,15 +9,21 @@ from plone.app.layout.navigation.navtree import buildFolderTree
 from plone.app.layout.navigation.root import getNavigationRoot
 from plone.app.portlets.browser import z3cformhelper
 from plone.app.portlets.portlets import navigation as basenav
+from plone.memoize import ram
 from plone.memoize.instance import memoize
+from plone.portlets.interfaces import IPortletManager
+from plone.portlets.interfaces import IPortletRetriever
+from plone.portlets.utils import hashPortletInfo
 from z3c.form import field
 from zope import schema
 from zope.component import adapts
 from zope.component import getMultiAdapter
+from zope.component import getUtilitiesFor
 from zope.interface import Interface
 from zope.interface import implements
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
+import time
 
 
 DEFAULT_SEARCH_BASE = 'search_base_context'
@@ -176,6 +183,24 @@ class Assignment(basenav.Assignment):
         super(Assignment, self).__init__(*args, **kwargs)
 
 
+def _render_cachekey(method, self):
+    # portlethash = hash(self.data)
+    portlethash = self.portlethash
+    # if self.expand_tree and self.search_base is not 'search_base_context':
+    #     # cannot cache on root, as long as selected class is set on template
+    #     path = get_root_path(self.context, self)
+    path = '/'.join(self.context.getPhysicalPath())
+    portal_membership = getToolByName(self.context, 'portal_membership')
+    timeout = time.time() // (60 * 60)  # cache for 60 min
+    cachekey = (
+        portlethash,
+        path,
+        portal_membership.getAuthenticatedMember().id,
+        timeout
+    )
+    return cachekey
+
+
 class Renderer(basenav.Renderer):
 
     @property
@@ -215,6 +240,29 @@ class Renderer(basenav.Renderer):
             query=queryBuilder(),
             strategy=strategy
         )
+
+    @property
+    @memoize
+    def portlethash(self):
+        portlethash = None
+        assignment = aq_base(self.data)
+
+        # Get the portlet info, to get the portlet hash.
+        # THIS IS CRAZY!
+        context = self.context
+        for name, manager in getUtilitiesFor(IPortletManager, context=context):
+            retriever = getMultiAdapter((context, manager), IPortletRetriever)
+            portlets = retriever.getPortlets()
+            for portlet in portlets:
+                if assignment == portlet['assignment']:
+                    # got you
+                    portlet['manager'] = self.manager.__name__  # not available in portlet info, yet. hurray.  # noqa
+                    portlethash = hashPortletInfo(portlet)
+        return portlethash
+
+    @ram.cache(_render_cachekey)
+    def render(self):
+        return self._template()
 
     _template = ViewPageTemplateFile('sitenavigation.pt')
     recurse = ViewPageTemplateFile('sitenavigation_recurse.pt')